requirements.txt
streamlit==1.40.0
langchain==0.3.7
langchain-openai==0.2.9
langchain-community==0.3.7
chromadb==0.5.20
pypdf==5.1.0
openai==1.55.0
python-dotenv==1.0.1
tiktoken==0.8.0
httpx<0.28.0
pdfplumber==0.11.4
